Ensure package sync integrity
Calculate a sha256 hash of the freshly built and synced package on both
the local and remote sides. If the hashes don't match, then attempt to
resync the package. If the hashes match, job completed. Rinse and repeat
up to five times, repeating only on failure.

We've found that scp can lie to us. It can return success even if it
didn't successfully transfer the file. The approach detailed above makes
our package building and syncing process more robust.